renamed:    user/client/src/Comm/CommomFuncAd.ts -> user/client/src/Comm/CommomFunc.ts 	modified:   user/client/src/Comm/Footer.tsx 	modified:   user/client/src/Pages/About/About.tsx 	modified:   user/client/src/Pages/Blog/Blog.tsx 	modified:   user/client/src/Pages/Blog/BlogDetail.tsx 	modified:   user/client/src/Pages/Coalition/Coalition.tsx 	modified:   user/client/src/Pages/Coalition/CoalitionAccount.tsx 	modified:   user/client/src/Pages/Contact/Contact.tsx 	modified:   user/client/src/Pages/Home/Home.tsx 	modified:   user/client/src/Pages/Hotel/Hotel.tsx 	modified:   user/client/src/Pages/Hotel/HotelDetail.tsx 	modified:   user/client/src/Pages/Tour/Tour.tsx 	modified:   user/client/src/Pages/Tour/TourDetail.tsx 	modified:   user/client/src/Pages/Tour/Tour_backup.tsx 	modified:   user/client/src/Test/Test.tsx
