+++ user/client/src/Comm/CommomFunc.ts
